federation-api: Improve claim_keys endpoint
