+++ crates/ruma-federation-api/CHANGELOG.md
@@ -1,5 +1,9 @@
 # [unreleased]
 
+Breaking Changes:
+
+* Change types of keys::claim_keys::v1 response to match the client-server endpoint
+
 Improvements:
 
 * Add master_keys and self_signing keys to keys::get_keys::v1 response
