+++ crates/ruma-federation-api/src/transactions/edu.rs
@@ -3,10 +3,10 @@
 use std::collections::BTreeMap;
 
 use js_int::UInt;
-use ruma_common::{encryption::DeviceKeys, presence::PresenceState};
-use ruma_events::{
-    from_raw_json_value, receipt::Receipt, room_key::RoomKeyToDeviceEventContent, EventType,
+use ruma_common::{
+    encryption::DeviceKeys, presence::PresenceState, to_device::DeviceIdOrAllDevices,
 };
+use ruma_events::{from_raw_json_value, receipt::Receipt, EventType};
 use ruma_identifiers::{DeviceIdBox, EventId, RoomId, UserId};
 use serde::{de, Deserialize, Serialize};
 use serde_json::{value::RawValue as RawJsonValue, Value as JsonValue};
@@ -258,7 +258,7 @@ pub struct DirectDeviceContent {
     /// The contents of the messages to be sent. These are arranged in a map
     /// of user IDs to a map of device IDs to message bodies. The device ID may
     /// also be *, meaning all known devices for the user.
-    pub messages: BTreeMap<UserId, BTreeMap<DeviceIdBox, RoomKeyToDeviceEventContent>>,
+    pub messages: BTreeMap<UserId, BTreeMap<DeviceIdOrAllDevices, Box<RawJsonValue>>>,
 }
 
 impl DirectDeviceContent {
